bring back support of aiml to python3 and other minor changes
- ask was broken due to moving of scripts
- ask was not available for python3
- aiml for python>=3.6 was not available(set to my repo, works fine as
tested  for now)
- same included in requirements.txt
- autopep8 and pycodestyle removed from requirements.txt, not needed
now, maybe downloaded by flake8
- aiml folder was renamed to brain for less conflict and appropriate
name
- brain class was moved to plugin/ask.py
- and some other changes, see in diff
