+++ jarviscli/plugins/brain/default.aiml
