+++ jarviscli/plugins/brain/std-startup.xml
