+++ requirements.txt
@@ -1,7 +1,6 @@
 aiml; python_version <= '2.7'
 appdirs==1.4.3
 archey4; python_version >= '3.5'
-autopep8
 beautifulsoup4==4.5.3
 bs4==0.0.1
 colorama==0.3.7
@@ -20,7 +19,6 @@ notify2
 packaging==16.8
 pick
 pluginmanager
-pycodestyle==2.3.1
 pycricbuzz
 pync==1.6.1
 pyowm==2.6.1
@@ -38,3 +36,4 @@ sympy
 termdown
 wikipedia
 youtube-dl
+https://github.com/appi147/pyaiml3/archive/master.zip ; python_version >= '3.5'
